direct intern string_view

+++ src/qjsp/object.hpp
@@ -12,9 +12,11 @@ namespace qjsp {
 struct Runtime;
 struct Context;
 
-using CFunction = Value(Context* ctx, Value this_val, int argc, const Value* argv);
+using CFunction = Value(Context *ctx, Value this_val, int argc, const Value *argv);
 
-struct Property { Value value = kUndefined; };
+struct Property {
+  Value value = kUndefined;
+};
 
 struct VarRef;
 
@@ -27,44 +29,44 @@ struct Object : GCObjectHeader {
   std::vector<Property> properties;
 
   // Closure data (only valid for bytecode_function class)
-  VarRef **var_refs       = nullptr;
-  int var_ref_count       = 0;
+  VarRef **var_refs = nullptr;
+  int var_ref_count = 0;
 
   // Class-specific data. Only valid for certain class_ids.
   struct CFunctionData {
-    void* realm = nullptr;
-    CFunction* fn = nullptr;
+    void *realm    = nullptr;
+    CFunction *fn  = nullptr;
     uint8_t length = 0;
-    int16_t magic = 0;
+    int16_t magic  = 0;
   };
   union {
     CFunctionData cfunc;
-    void* opaque;
+    void *opaque;
   } u{};
 
   // ── factories ────────────────────────────────────────────────────────
-  static Object* create(Runtime* rt, Object* proto, int class_id);
-  static Object* make_cfunc(Context* ctx, CFunction* fn, std::string_view name, int length);
+  static Object *create(Runtime *rt, Object *proto, uint16_t class_id);
+  static Object *make_cfunc(Context *ctx, CFunction *fn, std::string_view name, int length);
 
   // ── property access ──────────────────────────────────────────────────
   Value get_own(Atom atom) const;
-  bool set_own(Runtime* rt, Atom atom, Value value, int flags = kPropCWE);
-  bool define_own(Runtime* rt, Atom atom, Value value, int flags);
+  bool set_own(Runtime *rt, Atom atom, Value value, int flags = kPropCWE);
+  bool define_own(Runtime *rt, Atom atom, Value value, int flags);
   bool has_own(Atom atom) const { return shape && shape->find(atom) >= 0; }
   Value get(Atom atom) const;
 
-  void destroy(Runtime* rt);
+  void destroy(Runtime *rt);
 
   // ── GC ───────────────────────────────────────────────────────────────
-  void gc_mark(std::vector<GCObjectHeader*>& worklist);
+  void gc_mark(std::vector<GCObjectHeader *> &worklist);
 };
 
 // ── calling ───────────────────────────────────────────────────────────
 
-Value call(Context* ctx, Value func, Value this_val, int argc, const Value* argv);
+Value call(Context *ctx, Value func, Value this_val, int argc, const Value *argv);
 
 // ── global ─────────────────────────────────────────────────────────────
 
-void setup_global(Context* ctx, Object* global);
+void setup_global(Context *ctx, Object *global);
 
-}  // namespace qjsp
+} // namespace qjsp
